refactor: move user folder to data folder
