+++ src/infra/db/prisma/data/user/add-or-update-user-social-media-by-ids/add-or-update-user-social-media-by-ids-prisma-repo.spec.ts
@@ -2,7 +2,7 @@ import { type PrismaClient } from '@prisma/client'
 import { AddOrUpdateUserSocialMediaByIdsPrismaRepo } from './add-or-update-user-social-media-by-ids-prisma-repo'
 import { type SocialMediaModel } from '@/domain/models/social-media/social-media-model'
 import { PrismockClient } from 'prismock'
-import { PrismaHelper } from '../../helpers/prisma-helper'
+import { PrismaHelper } from '@/infra/db/prisma/helpers/prisma-helper'
 import MockDate from 'mockdate'
 import { type UserSocialMediaModel } from '@/domain/models/user/user-social-media-model'
 import { type UserModel } from '@/domain/models'
