+++ src/infra/db/prisma/data/user/add-user-preference/add-user-preference-prisma-repo.ts
@@ -1,5 +1,5 @@
 import { type AddUserPreferenceRepo } from '@/usecases/contracts/db/user'
-import { PrismaHelper } from '../../helpers/prisma-helper'
+import { PrismaHelper } from '@/infra/db/prisma/helpers/prisma-helper'
 import { type AddUserPreferenceData } from '@/domain/contracts/user/add-user-preference'
 
 export class AddUserPreferencePrismaRepo implements AddUserPreferenceRepo {
