+++ src/infra/db/prisma/data/user/find-preference-by-id/find-preference-by-id-prisma-repo.ts
@@ -1,14 +1,12 @@
 import { type PreferenceModel } from '@/domain/models'
-import { PrismaHelper } from '../../helpers/prisma-helper'
+import { PrismaHelper } from '@/infra/db/prisma/helpers/prisma-helper'
 import { type FindPreferenceByIdRepo } from '@/usecases/contracts/db/user'
 
 export class FindPreferenceByIdPrismaRepo implements FindPreferenceByIdRepo {
   async execute (userId: string): Promise<PreferenceModel | null> {
     const prisma = await PrismaHelper.getPrisma()
     const userPreferenceOrNull = await prisma.userPreference.findUnique({
-      where: {
-        id: userId
-      }
+      where: { id: userId }
     })
     return userPreferenceOrNull
   }
