+++ src/main/factories/infra/db/prisma/user/add-or-update-day-period-prisma-repo-factory.ts
@@ -1,4 +1,4 @@
-import { AddOrUpdateDayPeriodPrismaRepo } from '@/infra/db/prisma/user/add-or-update-day-period/add-or-update-day-period-prisma-repo'
+import { AddOrUpdateDayPeriodPrismaRepo } from '@/infra/db/prisma/data/user/add-or-update-day-period/add-or-update-day-period-prisma-repo'
 import { type AddOrUpdateDayPeriodRepo } from '@/usecases/contracts/db/user'
 
 export const makeAddOrUpdateDayPeriodPrismaRepo = (): AddOrUpdateDayPeriodRepo => {
