+++ src/main/factories/infra/db/prisma/user/add-or-update-game-place-prisma-repo-factory.ts
@@ -1,4 +1,4 @@
-import { AddOrUpdateGamePlacePrismaRepo } from '@/infra/db/prisma/user/add-or-update-game-place/add-or-update-game-place-prisma-repo'
+import { AddOrUpdateGamePlacePrismaRepo } from '@/infra/db/prisma/data/user/add-or-update-game-place/add-or-update-game-place-prisma-repo'
 
 export const makeAddOrUpdateGamePlacePrismaRepo = (): AddOrUpdateGamePlacePrismaRepo => {
   return new AddOrUpdateGamePlacePrismaRepo()
