+++ src/main/factories/infra/db/prisma/user/add-or-update-user-social-media-by-ids-repo-factory.ts
@@ -1,4 +1,4 @@
-import { AddOrUpdateUserSocialMediaByIdsPrismaRepo } from '@/infra/db/prisma/user/add-or-update-user-social-media-by-ids/add-or-update-user-social-media-by-ids-prisma-repo'
+import { AddOrUpdateUserSocialMediaByIdsPrismaRepo } from '@/infra/db/prisma/data/user/add-or-update-user-social-media-by-ids/add-or-update-user-social-media-by-ids-prisma-repo'
 import { type AddOrUpdateUserSocialMediaByIdsRepo } from '@/usecases/contracts/db/user/add-or-update-user-social-media-by-ids-repo'
 
 export const makeAddOrUpdateUserSocialMediaByIdsRepo = (): AddOrUpdateUserSocialMediaByIdsRepo => {
