+++ src/main/factories/infra/db/prisma/user/add-user-prisma-repo-factory.ts
@@ -1,4 +1,4 @@
-import { AddUserPrismaRepo } from '@/infra/db/prisma/user/add-user/add-user-prisma-repo'
+import { AddUserPrismaRepo } from '@/infra/db/prisma/data/user/add-user/add-user-prisma-repo'
 import { type AddUserRepo } from '@/usecases/contracts/db/user'
 
 export const makeAddUserPrismaRepo = (): AddUserRepo => {
