+++ src/main/factories/infra/db/prisma/user/find-user-by-email-prisma-repo-factory.ts
@@ -1,4 +1,4 @@
-import { FindUserByEmailPrismaRepo } from '@/infra/db/prisma/user/find-user-by-email/find-user-by-email-prisma-repo'
+import { FindUserByEmailPrismaRepo } from '@/infra/db/prisma/data/user/find-user-by-email/find-user-by-email-prisma-repo'
 import { type FindUserByEmailRepo } from '@/usecases/contracts/db/user'
 
 export const makeFindUserByEmailPrismaRepo = (): FindUserByEmailRepo => {
