+++ src/main/factories/infra/db/prisma/user/update-user-prisma-repo-factory.ts
@@ -1,4 +1,4 @@
-import { UpdateUserPrismaRepo } from '@/infra/db/prisma/user/update-user/update-user-prisma-repo'
+import { UpdateUserPrismaRepo } from '@/infra/db/prisma/data/user/update-user/update-user-prisma-repo'
 import { type UpdateUserRepo } from '@/usecases/contracts/db/user'
 
 export const makeUpdateUserPrismaRepo = (): UpdateUserRepo => {
